requirements.txt
# TrustMed AI - Requirements
# Install with: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Vector Database
chromadb>=0.4.0

# Embedding Models
sentence-transformers>=2.2.0

# Graph Database
neo4j>=5.0.0
langchain-neo4j>=0.1.0

# OpenAI Compatible API (for OpenRouter)
openai>=1.0.0

# HTTP Requests
requests>=2.28.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Optional: For UMLS API
# requests is already included above

# Image Processing & Multimodal
pillow>=10.0.0
opencv-python>=4.8.0
torch>=2.0.0
transformers>=4.30.0

# HuggingFace Datasets
datasets>=2.14.0


# Visualization
streamlit-agraph>=0.0.45
